Fix performance issues with outech-debug
We are now relying on the batching  mechanism of the robot.

We are not decimating the points anymore but we are sending them every 250ms grouped into 1 message.

Frontend will not split this message and update the react component once per message; it will instead will do a batch update.

The live components (graph & simulation) are not updated on every event change, but on a fixed interval (this will make sure 1second runs in 1second...)